refactor(examples): standardize C++ style and simplify logging
- Apply Google C++ naming conventions and reformat example code
- Replace ESP_LOGx with printf and remove esp_log.h and log tags
- Use esp_timer_get_time for timing independently of logging
- Standardize integer log formats and organize includes
- Retain #pragma once in example headers
- Migrate the LVGL demo from C to C++ and update build references
- Clean up static object lifetimes while preserving screen logging

+++ main/examples/lora_rx_sensitivity/lora_rx_sensitivity.cpp
@@ -2,12 +2,15 @@
  * @Description: 实现 LoRa 接收灵敏度测试会话、数据校验与丢包率统计
  * @Author: LILYGO_L
  * @Date: 2026-07-29 15:09:12
- * @LastEditTime: 2026-07-29 18:00:58
+ * @LastEditTime: 2026-09-22 17:04:01
  * @License: GPL 3.0
  */
 #include "lora_rx_sensitivity.h"
 
 #include <algorithm>
+#include <cinttypes>
+#include <cstddef>
+#include <cstdint>
 #include <cstdio>
 #include <cstring>
 
@@ -53,21 +56,21 @@ TestSession::TestSession(const char* radio_name, uint32_t frequency_hz)
     : radio_name_(radio_name), frequency_hz_(frequency_hz) {}
 
 void TestSession::PrintConfiguration() const {
-  const unsigned long bandwidth_khz =
-      (kUseHighFrequencyPath && std::strcmp(radio_name_, "LR1121") == 0)
-          ? 406UL
-          : (kUseHighFrequencyPath && std::strcmp(radio_name_, "LR2021") == 0)
-                ? 500UL : static_cast<unsigned long>(kBandwidthKhz);
-  printf("Frequency: %lu Hz\n", static_cast<unsigned long>(frequency_hz_));
-  printf("Modulation: LoRa, SF%u, BW %lu kHz, CR 4/5\n",
+  const uint32_t bandwidth_khz =
+      (kUseHighFrequencyPath && std::strcmp(radio_name_, "LR1121") == 0) ? 406UL
+      : (kUseHighFrequencyPath && std::strcmp(radio_name_, "LR2021") == 0)
+          ? 500UL
+          : static_cast<uint32_t>(kBandwidthKhz);
+  printf("Frequency: %" PRIu32 " Hz\n", static_cast<uint32_t>(frequency_hz_));
+  printf("Modulation: LoRa, SF%u, BW %" PRIu32 " kHz, CR 4/5\n",
       static_cast<unsigned int>(kSpreadingFactor), bandwidth_khz);
   printf("Packet: explicit header, preamble 8, payload %u bytes\n",
       static_cast<unsigned int>(kPayloadLength));
   printf("Payload CRC: on, IQ: standard, private sync word: 0x%02X\n",
       static_cast<unsigned int>(kSyncWord));
   printf("Low data rate optimization: on, receiver gain: boosted\n");
-  printf("Expected RF generator packet count: %lu\n",
-      static_cast<unsigned long>(kExpectedPacketCount));
+  printf("Expected RF generator packet count: %" PRIu32 "\n",
+      static_cast<uint32_t>(kExpectedPacketCount));
   printf("Pass criterion: PER <= %.1f%%\n", kPassingPerPercent);
   printf("RF generator data source: All 1\n");
 }
@@ -83,8 +86,8 @@ void TestSession::PrintSetup() const {
     printf("%02X%s", static_cast<unsigned int>(kExpectedPayload[index]),
         ((index + 1) % 16 == 0) ? "\n" : " ");
   }
-  printf("Receiver is ready. Transmit exactly %lu packets at each\n",
-      static_cast<unsigned long>(kExpectedPacketCount));
+  printf("Receiver is ready. Transmit exactly %" PRIu32 " packets at each\n",
+      static_cast<uint32_t>(kExpectedPacketCount));
   printf("input-power point; the first packet starts the test.\n");
   printf("Press BOOT once to clear statistics and restart the test.\n");
   printf("============================================================\n\n");
@@ -109,33 +112,28 @@ void TestSession::Start(uint32_t current_time_ms) {
   snr_ = {};
 
   printf("\n[PER] RF packet stream detected, test started\n");
-  printf("[PER] expecting exactly %lu packets\n",
-      static_cast<unsigned long>(kExpectedPacketCount));
+  printf("[PER] expecting exactly %" PRIu32 " packets\n",
+      static_cast<uint32_t>(kExpectedPacketCount));
 }
 
-void TestSession::Finish(
-    uint32_t current_time_ms, const char* reason) {
+void TestSession::Finish(uint32_t current_time_ms, const char* reason) {
   if (!running_) {
     return;
   }
   running_ = false;
-  if (std::strcmp(reason, "expected packet events reached") == 0) locked_ = true;
+  if (std::strcmp(reason, "expected packet events reached") == 0)
+    locked_ = true;
   last_progress_time_ms_ = current_time_ms;
   progress_timer_started_ = true;
 
   const uint32_t observed_bad = ObservedBadPackets();
   const uint32_t observed = good_packets_ + observed_bad;
-  const uint32_t bounded_observed =
-      std::min(observed, kExpectedPacketCount);
-  const uint32_t missing_packets =
-      kExpectedPacketCount - bounded_observed;
-  const uint32_t bounded_good =
-      std::min(good_packets_, kExpectedPacketCount);
-  const uint32_t failed_packets =
-      kExpectedPacketCount - bounded_good;
-  const double per_percent =
-      static_cast<double>(failed_packets) * 100.0 /
-      static_cast<double>(kExpectedPacketCount);
+  const uint32_t bounded_observed = std::min(observed, kExpectedPacketCount);
+  const uint32_t missing_packets = kExpectedPacketCount - bounded_observed;
+  const uint32_t bounded_good = std::min(good_packets_, kExpectedPacketCount);
+  const uint32_t failed_packets = kExpectedPacketCount - bounded_good;
+  const double per_percent = static_cast<double>(failed_packets) * 100.0 /
+                             static_cast<double>(kExpectedPacketCount);
   const uint32_t elapsed_ms = current_time_ms - start_time_ms_;
 
   printf("\n");
@@ -144,47 +142,40 @@ void TestSession::Finish(
   printf("Radio: %s\n", radio_name_);
   PrintConfiguration();
   printf("Finish reason: %s\n", reason);
-  printf("Elapsed: %lu ms\n", static_cast<unsigned long>(elapsed_ms));
-  printf("Expected packets: %lu\n",
-      static_cast<unsigned long>(kExpectedPacketCount));
-  printf("Correct packets: %lu\n",
-      static_cast<unsigned long>(bounded_good));
-  printf("Observed bad packets: %lu\n",
-      static_cast<unsigned long>(
-          std::min(observed_bad, kExpectedPacketCount)));
-  printf("Unobserved/missing packets: %lu\n",
-      static_cast<unsigned long>(missing_packets));
-  printf("Header errors: %lu\n",
-      static_cast<unsigned long>(header_errors_));
-  printf("CRC errors: %lu\n",
-      static_cast<unsigned long>(crc_errors_));
-  printf("Length errors: %lu\n",
-      static_cast<unsigned long>(length_errors_));
-  printf("Payload content errors: %lu\n",
-      static_cast<unsigned long>(payload_errors_));
-  printf("Packet read errors: %lu\n",
-      static_cast<unsigned long>(read_errors_));
-  printf("Driver/transport errors: %lu\n",
-      static_cast<unsigned long>(driver_errors_));
-  printf("Failed packets: %lu\n",
-      static_cast<unsigned long>(failed_packets));
+  printf("Elapsed: %" PRIu32 " ms\n", static_cast<uint32_t>(elapsed_ms));
+  printf("Expected packets: %" PRIu32 "\n",
+      static_cast<uint32_t>(kExpectedPacketCount));
+  printf("Correct packets: %" PRIu32 "\n", static_cast<uint32_t>(bounded_good));
+  printf("Observed bad packets: %" PRIu32 "\n",
+      static_cast<uint32_t>(std::min(observed_bad, kExpectedPacketCount)));
+  printf("Unobserved/missing packets: %" PRIu32 "\n",
+      static_cast<uint32_t>(missing_packets));
+  printf("Header errors: %" PRIu32 "\n", static_cast<uint32_t>(header_errors_));
+  printf("CRC errors: %" PRIu32 "\n", static_cast<uint32_t>(crc_errors_));
+  printf("Length errors: %" PRIu32 "\n", static_cast<uint32_t>(length_errors_));
+  printf("Payload content errors: %" PRIu32 "\n",
+      static_cast<uint32_t>(payload_errors_));
+  printf(
+      "Packet read errors: %" PRIu32 "\n", static_cast<uint32_t>(read_errors_));
+  printf("Driver/transport errors: %" PRIu32 "\n",
+      static_cast<uint32_t>(driver_errors_));
+  printf(
+      "Failed packets: %" PRIu32 "\n", static_cast<uint32_t>(failed_packets));
   printf("PER: %.3f%%\n", per_percent);
   printf("Result at this input level: %s\n",
       per_percent <= kPassingPerPercent ? "PASS" : "FAIL");
 
   if (packet_rssi_.count > 0) {
     printf("Packet RSSI min/avg/max: %.2f / %.2f / %.2f dBm\n",
-        packet_rssi_.minimum, packet_rssi_.Average(),
-        packet_rssi_.maximum);
+        packet_rssi_.minimum, packet_rssi_.Average(), packet_rssi_.maximum);
   }
   if (signal_rssi_.count > 0) {
     printf("Signal RSSI min/avg/max: %.2f / %.2f / %.2f dBm\n",
-        signal_rssi_.minimum, signal_rssi_.Average(),
-        signal_rssi_.maximum);
+        signal_rssi_.minimum, signal_rssi_.Average(), signal_rssi_.maximum);
   }
   if (snr_.count > 0) {
-    printf("SNR min/avg/max: %.2f / %.2f / %.2f dB\n",
-        snr_.minimum, snr_.Average(), snr_.maximum);
+    printf("SNR min/avg/max: %.2f / %.2f / %.2f dB\n", snr_.minimum,
+        snr_.Average(), snr_.maximum);
   }
 
   printf("Sensitivity is the lowest calibrated DUT input level that\n");
@@ -219,14 +210,12 @@ void TestSession::Poll(uint32_t current_time_ms) {
     return;
   }
 
-  if ((current_time_ms - last_activity_time_ms_) >=
-      kIdleFinishTimeMs) {
+  if ((current_time_ms - last_activity_time_ms_) >= kIdleFinishTimeMs) {
     Finish(current_time_ms, "RF packet stream ended");
     return;
   }
 
-  if ((current_time_ms - last_progress_time_ms_) >=
-      kProgressPrintIntervalMs) {
+  if ((current_time_ms - last_progress_time_ms_) >= kProgressPrintIntervalMs) {
     PrintProgress(current_time_ms);
     last_progress_time_ms_ = current_time_ms;
   }
@@ -252,13 +241,12 @@ void TestSession::Restart(uint32_t current_time_ms) {
   snr_ = {};
 
   printf("\n[PER] test restarted by BOOT button\n");
-  printf("[PER] waiting for a new %lu-packet RF stream\n",
-      static_cast<unsigned long>(kExpectedPacketCount));
+  printf("[PER] waiting for a new %" PRIu32 "-packet RF stream\n",
+      static_cast<uint32_t>(kExpectedPacketCount));
 }
 
-void TestSession::RecordPacket(const uint8_t* payload,
-    size_t payload_length, const PacketMetrics& metrics,
-    uint32_t current_time_ms) {
+void TestSession::RecordPacket(const uint8_t* payload, size_t payload_length,
+    const PacketMetrics& metrics, uint32_t current_time_ms) {
   if (!running_) {
     Start(current_time_ms);
   }
@@ -267,9 +255,8 @@ void TestSession::RecordPacket(const uint8_t* payload,
   last_activity_time_ms_ = current_time_ms;
   if (payload == nullptr || payload_length != kExpectedPayload.size()) {
     ++payload_errors_;
-  } else if (std::memcmp(
-                 payload, kExpectedPayload.data(), kExpectedPayload.size()) !=
-             0) {
+  } else if (std::memcmp(payload, kExpectedPayload.data(),
+                 kExpectedPayload.size()) != 0) {
     ++payload_errors_;
   } else {
     ++good_packets_;
@@ -308,9 +295,9 @@ void TestSession::RecordPacketError(
 
   const uint32_t observed = ObservedPackets();
   if (observed <= 10) {
-    printf("[PER] observed %s error, packet event %lu/%lu\n",
-        PacketErrorToString(error), static_cast<unsigned long>(observed),
-        static_cast<unsigned long>(kExpectedPacketCount));
+    printf("[PER] observed %s error, packet event %" PRIu32 "/%" PRIu32 "\n",
+        PacketErrorToString(error), static_cast<uint32_t>(observed),
+        static_cast<uint32_t>(kExpectedPacketCount));
   }
   CheckForCompletion(current_time_ms);
 }
@@ -327,59 +314,50 @@ void TestSession::PrintProgress(uint32_t current_time_ms) const {
   const uint32_t remaining_packets =
       kExpectedPacketCount - std::min(observed, kExpectedPacketCount);
   const uint32_t elapsed_ms = current_time_ms - start_time_ms_;
-  const double completion_percent =
-      static_cast<double>(observed) * 100.0 /
-      static_cast<double>(kExpectedPacketCount);
+  const double completion_percent = static_cast<double>(observed) * 100.0 /
+                                    static_cast<double>(kExpectedPacketCount);
   const double observed_error_percent =
-      observed == 0
-          ? 0.0
-          : static_cast<double>(observed_bad) * 100.0 /
-                static_cast<double>(observed);
+      observed == 0 ? 0.0
+                    : static_cast<double>(observed_bad) * 100.0 /
+                          static_cast<double>(observed);
 
   printf("\n");
   printf("============================================================\n");
   printf("[PER] INTERIM RESULT\n");
   printf("Radio: %s\n", radio_name_);
-  printf("Elapsed: %lu ms\n", static_cast<unsigned long>(elapsed_ms));
-  printf("Expected packets: %lu\n",
-      static_cast<unsigned long>(kExpectedPacketCount));
-  printf("Observed packets: %lu\n",
-      static_cast<unsigned long>(observed));
-  printf("Remaining to expected count: %lu\n",
-      static_cast<unsigned long>(remaining_packets));
-  printf("Correct packets: %lu\n",
-      static_cast<unsigned long>(good_packets_));
-  printf("Observed bad packets: %lu\n",
-      static_cast<unsigned long>(observed_bad));
-  printf("Header errors: %lu\n",
-      static_cast<unsigned long>(header_errors_));
-  printf("CRC errors: %lu\n",
-      static_cast<unsigned long>(crc_errors_));
-  printf("Length errors: %lu\n",
-      static_cast<unsigned long>(length_errors_));
-  printf("Payload content errors: %lu\n",
-      static_cast<unsigned long>(payload_errors_));
-  printf("Packet read errors: %lu\n",
-      static_cast<unsigned long>(read_errors_));
-  printf("Driver/transport errors: %lu\n",
-      static_cast<unsigned long>(driver_errors_));
+  printf("Elapsed: %" PRIu32 " ms\n", static_cast<uint32_t>(elapsed_ms));
+  printf("Expected packets: %" PRIu32 "\n",
+      static_cast<uint32_t>(kExpectedPacketCount));
+  printf("Observed packets: %" PRIu32 "\n", static_cast<uint32_t>(observed));
+  printf("Remaining to expected count: %" PRIu32 "\n",
+      static_cast<uint32_t>(remaining_packets));
+  printf(
+      "Correct packets: %" PRIu32 "\n", static_cast<uint32_t>(good_packets_));
+  printf("Observed bad packets: %" PRIu32 "\n",
+      static_cast<uint32_t>(observed_bad));
+  printf("Header errors: %" PRIu32 "\n", static_cast<uint32_t>(header_errors_));
+  printf("CRC errors: %" PRIu32 "\n", static_cast<uint32_t>(crc_errors_));
+  printf("Length errors: %" PRIu32 "\n", static_cast<uint32_t>(length_errors_));
+  printf("Payload content errors: %" PRIu32 "\n",
+      static_cast<uint32_t>(payload_errors_));
+  printf(
+      "Packet read errors: %" PRIu32 "\n", static_cast<uint32_t>(read_errors_));
+  printf("Driver/transport errors: %" PRIu32 "\n",
+      static_cast<uint32_t>(driver_errors_));
   printf("Completion: %.2f%%\n", completion_percent);
-  printf("Observed packet error ratio: %.3f%%\n",
-      observed_error_percent);
+  printf("Observed packet error ratio: %.3f%%\n", observed_error_percent);
 
   if (packet_rssi_.count > 0) {
     printf("Packet RSSI min/avg/max: %.2f / %.2f / %.2f dBm\n",
-        packet_rssi_.minimum, packet_rssi_.Average(),
-        packet_rssi_.maximum);
+        packet_rssi_.minimum, packet_rssi_.Average(), packet_rssi_.maximum);
   }
   if (signal_rssi_.count > 0) {
     printf("Signal RSSI min/avg/max: %.2f / %.2f / %.2f dBm\n",
-        signal_rssi_.minimum, signal_rssi_.Average(),
-        signal_rssi_.maximum);
+        signal_rssi_.minimum, signal_rssi_.Average(), signal_rssi_.maximum);
   }
   if (snr_.count > 0) {
-    printf("SNR min/avg/max: %.2f / %.2f / %.2f dB\n",
-        snr_.minimum, snr_.Average(), snr_.maximum);
+    printf("SNR min/avg/max: %.2f / %.2f / %.2f dB\n", snr_.minimum,
+        snr_.Average(), snr_.maximum);
   }
 
   printf("Final PER is calculated only after the test ends.\n");
@@ -393,8 +371,8 @@ void TestSession::CheckForCompletion(uint32_t current_time_ms) {
 }
 
 uint32_t TestSession::ObservedBadPackets() const {
-  return header_errors_ + crc_errors_ + length_errors_ +
-         payload_errors_ + read_errors_;
+  return header_errors_ + crc_errors_ + length_errors_ + payload_errors_ +
+         read_errors_;
 }
 
 uint32_t TestSession::ObservedPackets() const {
